+++ src/applications/edu-benefits/10203/content/benefitSelection.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 
 export const benefitsLabels = {
   chapter33: (
@@ -114,4 +115,26 @@ export const survivorBenefitsLabels = {
       </a>
     </p>
   ),
+};
+
+export const ineligibleAlert = () => {
+  const text = (
+    <div>
+      <p>
+        The Rogers STEM Scholarship is only for Post-9/11 Gi Bill and Fry
+        Scholarship recipients.
+      </p>
+      <p>
+        If you think you may be eligible, you can still choose to apply for the
+        Rogers STEM Scholarship
+      </p>
+    </div>
+  );
+  return (
+    <AlertBox
+      status="warning"
+      headline="You may not be eligible"
+      content={text}
+    />
+  );
 };
