+++ src/applications/edu-benefits/10203/helpers.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 
 const isEdithNourseRogersScholarship = form =>
   form.isEdithNourseRogersScholarship;
@@ -26,24 +25,6 @@ export const display10203StemFlow = form =>
   (isEligibleForEdithNourseRogersScholarship(form) ||
     determineEligibilityFor10203Stem(form));
 
-export const MGIBAlert = () => {
-  const text = (
-    <div>
-      <p>
-        The Rogers STEM Scholarship is only for Post-9/11 Gi Bill and Fry
-        Scholarship recipients.
-      </p>
-      <p>
-        If you think you may be eligible, you can still choose to apply for the
-        Rogers STEM Scholarship
-      </p>
-    </div>
-  );
-  return (
-    <AlertBox
-      status="warning"
-      headline="You may not be eligible"
-      content={text}
-    />
-  );
-};
+export const displayIneligibleAlert = form =>
+  form['view:benefit'].chapter30 === undefined ||
+  form['view:benefit'].chapter30 === false;
